+++ pvinspect/__init__.py
@@ -2,4 +2,4 @@ from . import common, data, integration, preproc
 
 __all__ = ["data", "preproc", "common", "integration"]
 
-__version__ = "0.4.1"
+__version__ = "0.4.2"